Create a new function : playGame()
Moved global variables and function calls inside of it.
A loop will run 5 times.
At each iteration, two variable will store choices of human and computer respectively.

I made some changes in playGround() function.
When human wins, a "human" string will be returned.
When computer wins, a "computer" string is returned.

A 'winner' variable will store the result of playGround().
Following if-else blocks will humanScore or computerScore by 1 depending on who is the winner.

After the loop is over, humanScore and computerScore will be compared and result will be declared in the following if-else blocks.