Display winner
Display the winner when any player's score reaches to 5.
Logic for this is written in playRound(). New html div element will be created on the spot with the class of 'winner'. Text of declaration of winner will be added in winner div. After all of that, the newly created winner div will be appended to the html body.