Create restart button
A restart button will automatically created once any player's score reaches 5.

Firstly, a restart div is created and a simple text indicating that the game is over and informing player with option to restart the game.

If player decides to click the restart button, everything will be restart as if player opened the page for the first time.